requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
